+++ app/code/local/Stuntcoders/FbPixel/Model/Observer.php
@@ -7,6 +7,12 @@ class Stuntcoders_FbPixel_Model_Observer
         Mage::getSingleton('customer/session')->setIsProductAddedToWishlist(true);
     }
 
+    public function onAddToCart($observer)
+    {
+        $product = $observer->getProduct();
+        Mage::getSingleton('customer/session')->setAddedProductName($product->getName());
+    }
+
     public function onCompleteRegistration()
     {
         Mage::getSingleton('customer/session')->setCompleteRegistration(true);