+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 ### ADDED:
 - Add to wishlist event
+- View content event
 
 ## 0.1.2 - 2016-03-15
 ### CHANGED: