+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### ADDED:
 - Add to wishlist event
 - View content event
+- Initiate checkout event
 
 ## 0.1.2 - 2016-03-15
 ### CHANGED: