+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - View content event
 - Initiate checkout event
 - Complete registration event
+- Add to cart event
 
 ## 0.1.2 - 2016-03-15
 ### CHANGED: